+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Contributors (alphabetical): Matthias van der hallen, Pierre Carbonnelle.
 * fix #90: error in `sum` aggregate
 
 ### Interactive Consultant
+* Merge #85: improve performance of IC (better use of cache)
 * fix #89: `constructed from` types create error in IC